Fixed tests running parallel
